Adopting TFS commits from 2661361 to 5232c61 (gpt-6-astra)

+++ core/src/main/java/greycos/solver/core/impl/bavet/common/tuple/TupleLifecycle.java
@@ -73,7 +73,7 @@ public interface TupleLifecycle<Tuple_ extends Tuple> extends ActivitySupport {
 
   static <Stream_ extends BavetStream, Tuple_ extends Tuple> TupleLifecycle<Tuple_> profiling(
       InnerConstraintProfiler constraintProfiler,
-      long lifecycleId,
+      int lifecycleId,
       Stream_ stream,
       TupleLifecycle<Tuple_> delegate) {
     if (delegate instanceof AggregatedTupleLifecycle) {
