+++ core/src/main/java/greycos/solver/core/impl/score/stream/bavet/common/ConstraintNodeBuildHelper.java
@@ -49,7 +49,7 @@ public final class ConstraintNodeBuildHelper<Solution_, Score_ extends Score<Sco
   private final Map<BavetScoringConstraintStream<Solution_>, Scorer<?>> streamToScorers =
       new HashMap<>();
 
-  private long nextLifecycleProfilingId = 0;
+  private int nextLifecycleProfilingId = 0;
 
   public ConstraintNodeBuildHelper(
       ConsistencyTracker<Solution_> consistencyTracker,
